requirements.txt
flask==3.0.0
pymysql==1.1.0
sqlalchemy==1.4.46
requests==2.31.0
python-dotenv==1.0.0
google-generativeai==0.3.2
slack-sdk==3.26.0